+++ R/get_advanced_match_stats.R
@@ -1,3 +1,43 @@
+#' @importFrom xml2 xml_find_all xml_attr xml_text
+#' @importFrom dplyr mutate
+#' @importFrom stats setNames
+.add_player_href <- function(df, parent_element, player_xpath) {
+  player_elements <- xml2::xml_find_all(parent_element, player_xpath)
+  players <- stats::setNames(
+    xml2::xml_attr(player_elements, "href"),
+    xml2::xml_text(player_elements)
+  )
+  res <- dplyr::mutate(
+    df,
+    "Player_Href" = players[df$Player],
+    .after = "Player"
+  )
+  return(res)
+}
+
+#' @importFrom rvest html_nodes html_text html_table
+#' @importFrom purrr pluck
+.extract_team_players <- function(match_page, xml_elements, team_idx, home_away) {
+
+  team <- match_page %>%
+    rvest::html_nodes("div+ strong a") %>%
+    rvest::html_text() %>%
+    purrr::pluck(team_idx)
+
+  team_stat <- xml_elements[team_idx] %>%
+    rvest::html_table() %>%
+    data.frame() %>%
+    .clean_match_advanced_stats_data()
+
+  team_stat <- .add_player_href(
+    team_stat,
+    parent_element = xml_elements[team_idx],
+    player_xpath = ".//tbody/tr/th/a"
+  )
+  res <- cbind(list("Team" = team, "Home_Away" = home_away), team_stat)
+  return(res)
+}
+
 #' Get FBref advanced match stats
 #'
 #' Returns data frame of selected statistics for each match, for either whole team or individual players.
@@ -84,40 +124,22 @@ fb_advanced_match_stats <- function(match_url, stat_type, team_or_player, time_p
           rvest::html_nodes("table")
 
       }
-      # shots stat type is not yet built in to the function
-      # else if(stat_type == "shots") {
-      #   stat_df <- all_tables[which(stringr::str_detect(all_tables %>% rvest::html_attr("id"), "all_shots"))] %>% .[1] %>%
-      #     rvest::html_table()
-      #
-      # }
-
-
 
       if(length(stat_df) != 0) {
 
         if(!stat_type %in% c("shots")) {
-          Team <- match_page %>%
-            rvest::html_nodes("div+ strong a") %>%
-            rvest::html_text() %>% .[1]
-
-          home_stat <- stat_df[1] %>%
-            rvest::html_table() %>% data.frame() %>%
-            .clean_match_advanced_stats_data()
-
-          Home_Away <- "Home"
-          home_stat <- cbind(Team, Home_Away, home_stat)
-
-          Team <- match_page %>%
-            rvest::html_nodes("div+ strong a") %>%
-            rvest::html_text() %>% .[2]
-
-          away_stat <- stat_df[2] %>%
-            rvest::html_table() %>% data.frame() %>%
-            .clean_match_advanced_stats_data()
-
-          Home_Away <- "Away"
-          away_stat <- cbind(Team, Home_Away, away_stat)
-
+          home_stat <- .extract_team_players(
+            match_page = match_page,
+            xml_elements = stat_df,
+            team_idx = 1,
+            home_away = "Home"
+          )
+          away_stat <- .extract_team_players(
+            match_page = match_page,
+            xml_elements = stat_df,
+            team_idx = 2,
+            home_away = "Away"
+          )
           stat_df_output <- dplyr::bind_rows(home_stat, away_stat)
 
           if(any(grepl("Nation", colnames(stat_df_output)))) {
@@ -174,7 +196,8 @@ fb_advanced_match_stats <- function(match_url, stat_type, team_or_player, time_p
 
   suppressWarnings(
     final_df <- match_url %>%
-      purrr::map_df(get_each_match_statistic) )
+      purrr::map_df(get_each_match_statistic)
+  )
 
   return(final_df)
 }