+++ tests/testthat/test-fbref.R
@@ -1,5 +1,5 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
- #======================================= IMPORTANT: ======================================#
+#======================================= IMPORTANT: ======================================#
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
 
 # as at 2022-07-08:
@@ -74,8 +74,10 @@ test_that("fb_advanced_match_stats() works", {
   #                                      stat_type = "keeper", team_or_player = "player", time_pause = 4), "list")
 
   # test that multiple match_url can be passed to the function
-  test_urls <- c("https://fbref.com/en/matches/e0a84e7e/Southampton-Manchester-United-November-29-2020-Premier-League",
-                 "https://fbref.com/en/matches/703db983/Paris-Saint-Germain-Toulouse-August-25-2019-Ligue-1")
+  test_urls <- c(
+    "https://fbref.com/en/matches/e0a84e7e/Southampton-Manchester-United-November-29-2020-Premier-League",
+    "https://fbref.com/en/matches/703db983/Paris-Saint-Germain-Toulouse-August-25-2019-Ligue-1"
+  )
   test_df <- fb_advanced_match_stats(match_url = test_urls, stat_type = "possession", team_or_player = "team", time_pause = 4)
   expect_type(test_df, "list")
   expect_equal(nrow(test_df), 4)
@@ -84,8 +86,14 @@ test_that("fb_advanced_match_stats() works", {
   # expect_error(fb_advanced_match_stats(match_url = "aaa.aaa"))
 
   # test that an invalid stat_type will error
-  expect_error(fb_advanced_match_stats(match_url = "https://fbref.com/en/matches/c0996cac/Bordeaux-Nantes-August-21-2020-Ligue-1",
-                                        stat_type = "test", team_or_player = "team", time_pause = 4))
+  expect_error(
+    fb_advanced_match_stats(
+      match_url = "https://fbref.com/en/matches/c0996cac/Bordeaux-Nantes-August-21-2020-Ligue-1",
+      stat_type = "test",
+      team_or_player = "team",
+      time_pause = 4
+    )
+  )
 
 })
 
@@ -242,8 +250,10 @@ test_that("fb_big5_advanced_season_stats() works", {
 Sys.sleep(3)
 test_that("fb_match_shooting() works", {
   testthat::skip_on_cran()
-  test_urls_multiple <- c("https://fbref.com/en/matches/c0996cac/Bordeaux-Nantes-August-21-2020-Ligue-1",
-                          "https://fbref.com/en/matches/9cbccb37/Dijon-Angers-August-22-2020-Ligue-1")
+  test_urls_multiple <- c(
+    "https://fbref.com/en/matches/c0996cac/Bordeaux-Nantes-August-21-2020-Ligue-1",
+    "https://fbref.com/en/matches/9cbccb37/Dijon-Angers-August-22-2020-Ligue-1"
+  )
   shot_multiple_matches <- fb_match_shooting(test_urls_multiple, time_pause = 4)
   expect_type(shot_multiple_matches, "list")
   expect_false(nrow(shot_multiple_matches) == 0)
@@ -281,9 +291,14 @@ test_that("fb_team_match_results() works", {
 Sys.sleep(3)
 test_that("fb_player_season_stats() works", {
   testthat::skip_on_cran()
-  multiple_playing_time <- fb_player_season_stats(player_url = c("https://fbref.com/en/players/d70ce98e/Lionel-Messi",
-                                                                 "https://fbref.com/en/players/dea698d9/Cristiano-Ronaldo"),
-                                                  stat_type = "playing_time", time_pause = 4)
+  multiple_playing_time <- fb_player_season_stats(
+    player_url = c(
+      "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
+      "https://fbref.com/en/players/dea698d9/Cristiano-Ronaldo"
+    ),
+    stat_type = "playing_time",
+    time_pause = 4
+  )
   expect_type(multiple_playing_time, "list")
   expect_false(nrow(multiple_playing_time) == 0)
 
@@ -297,10 +312,14 @@ test_that("fb_player_season_stats() works", {
   expect_type(multiple_national_standard_stats, "list")
   expect_false(nrow(multiple_national_standard_stats) == 0)
 
-  expect_no_error(victor_ulloa_national_stats <- fb_player_season_stats(
-    player_url = "https://fbref.com/en/players/e7056f05/Victor-Ulloa",
-    stat_type = "standard", time_pause = 4, national = TRUE
-  ))
+  expect_no_error(
+    victor_ulloa_national_stats <- fb_player_season_stats(
+      player_url = "https://fbref.com/en/players/e7056f05/Victor-Ulloa",
+      stat_type = "standard",
+      time_pause = 4,
+      national = TRUE
+    )
+  )
   expect_type(victor_ulloa_national_stats, "list")
   expect_true(nrow(victor_ulloa_national_stats) == 0)
 
@@ -381,19 +400,19 @@ Sys.sleep(3)
 test_that("fb_league_stats() for players works", {
   testthat::skip_on_cran()
   testthat::skip_on_ci()
-  expected_player_shooting_cols <- c("Rk", "Player", "Nation", "Pos", "Squad", "Age", "Born", "Mins_Per_90", "Gls", "Sh_Standard", "SoT_Standard", "SoT_percent_Standard", "Sh_per_90_Standard", "SoT_per_90_Standard", "G_per_Sh_Standard", "G_per_SoT_Standard", "Dist_Standard", "FK_Standard", "PK", "PKatt", "xG_Expected", "npxG_Expected", "npxG_per_Sh_Expected", "G_minus_xG_Expected", "np:G_minus_xG_Expected", "Matches", "url")
+  expected_player_shooting_cols <- c("Rk", "Player", "Player_Href", "Nation", "Pos", "Squad", "Age", "Born", "Mins_Per_90", "Gls_Standard", "Sh_Standard", "SoT_Standard", "SoT_percent_Standard", "Sh_per_90_Standard", "SoT_per_90_Standard", "G_per_Sh_Standard", "G_per_SoT_Standard", "Dist_Standard", "FK_Standard", "PK_Standard", "PKatt_Standard", "xG_Expected", "npxG_Expected", "npxG_per_Sh_Expected", "G_minus_xG_Expected", "np:G_minus_xG_Expected", "Matches", "url")
   epl_player_shooting_22 <- fb_league_stats(
     country = "BRA",
     gender = "M",
     season_end_year = 2022,
     tier = "1st",
-    stat_type = "standard",
+    stat_type = "shooting",
     team_or_player = "player"
   )
   expect_gt(nrow(epl_player_shooting_22), 0)
   expect_setequal(colnames(epl_player_shooting_22), expected_player_shooting_cols)
 
-  expected_player_misc_cols <- c("Rk", "Player", "Nation", "Pos", "Squad", "Age", "Born", "Mins_Per_90", "CrdY", "CrdR", "2CrdY", "Fls", "Fld", "Off", "Crs", "Int", "TklW", "PKwon", "PKcon", "OG", "Recov", "Won_Aerial Duels", "Lost_Aerial Duels", "Won_percent_Aerial Duels", "Matches", "url")
+  expected_player_misc_cols <- c("Rk", "Player", "Player_Href", "Nation", "Pos", "Squad", "Age", "Born", "Mins_Per_90", "CrdY", "CrdR", "2CrdY", "Fls", "Fld", "Off", "Crs", "Int", "TklW", "PKwon", "PKcon", "OG", "Recov", "Won_Aerial Duels", "Lost_Aerial Duels", "Won_percent_Aerial Duels", "Matches", "url")
   ## testing a lot would take too long, so just test multiple years since that is the most likely input param to have multiple values
   multi_season_player_misc <- fb_league_stats(
     # country = c("ITA", "ESP"),